+++ apps/web-giddh/src/app/billing-details/billingDetail.component.ts
@@ -87,7 +87,7 @@ export class BillingDetailComponent implements OnInit, OnDestroy, AfterViewInit 
     /* This will hold common JSON data */
     public commonLocaleData: any = {};
     /** control for the MatSelect filter keyword */
-    public searchBillingStates: FormControl = new FormControl();
+    public searchBillingStates: string = "";
     /** Billing States list */
     public filteredBillingStates: IOption[] = [];
     /** control for the MatSelect filter keyword */
@@ -124,22 +124,15 @@ export class BillingDetailComponent implements OnInit, OnDestroy, AfterViewInit 
     }
 
     public ngOnInit(): void {
-
-        /** This will be use for reset state value if input is empty */
-        this.store.pipe(select(s => s.general.states), takeUntil(this.destroyed$)).subscribe(res => {
-            if (res) {
-                Object.keys(res.stateList).forEach(key => {
-                    if (res.stateList[key]?.code === this.createNewCompany?.addresses[0]?.stateCode) {
-                        this.checkAndResetValue(this.searchBillingStates, res.stateList[key].name);
-                    }
-                });
-            }
-        });
-
-        /** This will use for filter states  */
-        this.searchBillingStates.valueChanges.pipe(takeUntil(this.destroyed$)).subscribe(search => {
-            this.filterStates(search);
-        });
+        /* RAZORPAY */
+        if (window['Razorpay'] === undefined) {
+            let scriptTag = document.createElement('script');
+            scriptTag.src = 'https://checkout.razorpay.com/v1/checkout.js';
+            scriptTag.type = 'text/javascript';
+            scriptTag.defer = true;
+            document.body.appendChild(scriptTag);
+        }
+        /* RAZORPAY */
 
         /** This will use for filter country  */
         this.searchCountry.valueChanges.pipe(takeUntil(this.destroyed$)).subscribe(search => {
@@ -155,9 +148,8 @@ export class BillingDetailComponent implements OnInit, OnDestroy, AfterViewInit 
                 this.showGstAndTaxUsingCountryName(activeCompany.countryV2?.countryName);
                 // this.searchCountry.setValue({ label: activeCompany.countryV2?.countryName });
                 this.activeCompany = activeCompany;
-                this.getStates();
-                // this.getCountry();
                 this.reFillForm();
+                this.getStates();
             }
         });
 
@@ -375,7 +367,7 @@ export class BillingDetailComponent implements OnInit, OnDestroy, AfterViewInit 
      * @memberof BillingDetailComponent
      */
     public onStateChange(event: any): void {
-        this.billingDetailsObj.stateCode = event?.option?.value?.value;
+        this.billingDetailsObj.stateCode = event?.value;
         this.cdRef.detectChanges();
     }
 
@@ -425,7 +417,7 @@ export class BillingDetailComponent implements OnInit, OnDestroy, AfterViewInit 
 
         };
         setTimeout(() => {
-            this.razorpay = new (window as any).Razorpay(this.options);
+            this.razorpay = new window['Razorpay'](this.options);
         }, 1000);
     }
 
@@ -489,6 +481,7 @@ export class BillingDetailComponent implements OnInit, OnDestroy, AfterViewInit 
                 this.countrySource$ = observableOf(this.countrySource);
                 setTimeout(() => {
                     this.showLoader = false;
+                    this.cdRef.detectChanges();
                 }, 3000);
             } else {
                 let countryRequest = new CountryRequest();
@@ -519,7 +512,7 @@ export class BillingDetailComponent implements OnInit, OnDestroy, AfterViewInit 
 
                     if (this.createNewCompany !== undefined && this.createNewCompany.addresses !== undefined && this.createNewCompany.addresses[0] !== undefined) {
                         if (res.stateList[key].code === this.createNewCompany.addresses[0].stateCode) {
-                            this.searchBillingStates.setValue({ label: res.stateList[key].name });
+                            this.searchBillingStates = res.stateList[key].name;
                             this.selectedState = res.stateList[key].name;
                             this.billingDetailsObj.stateCode = res.stateList[key].code;
                         }
@@ -536,6 +529,7 @@ export class BillingDetailComponent implements OnInit, OnDestroy, AfterViewInit 
                 statesRequest.country = this.createNewCompany ? this.createNewCompany.country : this.activeCompany.countryV2 ? this.activeCompany.countryV2.alpha2CountryCode : '';
                 this.store.dispatch(this.generalActions.getAllState(statesRequest));
             }
+            this.cdRef.detectChanges();
         });
     }
 
@@ -661,11 +655,6 @@ export class BillingDetailComponent implements OnInit, OnDestroy, AfterViewInit 
     public resetValueIfOptionNotSelected(field: string): void {
         setTimeout(() => {
             switch (field) {
-                case "billingState":
-                    const stateObj = this.companyStatesSource?.filter(state => state.value === this.billingDetailsObj.stateCode);
-                    this.checkAndResetValue(this.searchBillingStates, stateObj[0]?.label);
-                    break;
-
                 case "billingCountry":
                     this.checkAndResetValue(this.searchCountry, this.activeCompany.country);
                     break;
@@ -713,7 +702,7 @@ export class BillingDetailComponent implements OnInit, OnDestroy, AfterViewInit 
      * @memberof BillingDetailComponent
      */
     public onSelectChangeCountry(evt: any) {
-        this.searchBillingStates.setValue('');
+        this.searchBillingStates = '';
         this.getUpdatedStateCodes(evt.source.value.value, true);
         if (evt.source.value.label === 'India' && this.activeCompany?.country === 'India') {
             this.showGstinNo = true;
